Created getInvestmentAccount API which includes all the methods requi…

+++ src/main/java/com/gee/repository/InvestmentAccountRepository.java
@@ -4,7 +4,7 @@ import com.gee.model.InvestmentAccount;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 //@Repository > not needed, done by default for us
-public interface InvestmentAccountRepository extends JpaRepository<InvestmentAccount,Long> {
+public interface InvestmentAccountRepository extends JpaRepository<InvestmentAccount,Integer> {
